Moved router to components folder to separate animations.
